+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [Report diagnostics in `Type::from_ast`](https://github.com/amp-lang/ampc/issues/5) (report invalid types)
 - Report type mismatches in `const` declarations
 - [Implement using function calls as values](https://github.com/amp-lang/ampc/issues/8)
+- [Implement implicit returns](https://github.com/amp-lang/ampc/issues/6)
 
 ### [`snapshot-02-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-02-06-2023-a)
 (initial snapshot)